+++ Code/LYRUIParticipantSectionHeaderView.h
@@ -12,13 +12,7 @@
  @abstract The `LYRUIParticipantSectionHeaderView` displays a letter representing a section
  in the participant picker.
  */
-@interface LYRUIParticipantSectionHeaderView : UIView
-
-/**
- @abstract Tells the receiver which letter to display for a given section.
- @param key The letter to display.
- */
-- (id)initWithKey:(NSString *)key;
+@interface LYRUIParticipantSectionHeaderView : UITableViewHeaderFooterView
 
 /**
  @abstract The label representing the section of the participant picker.